Refactoring property
