+++ src/main/kotlin/io/github/siyual_park/property/TokenProperty.kt
@@ -1,10 +1,10 @@
-package io.github.siyual_park.confg
+package io.github.siyual_park.property
 
 import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.boot.context.properties.ConstructorBinding
 
 @ConstructorBinding
-@ConfigurationProperties(prefix = "app.token")
+@ConfigurationProperties(prefix = "application.token")
 data class TokenProperty(
     val secret: String,
     val expiresIn: Long
