+++ src/main/resources/application.properties
@@ -6,5 +6,5 @@ spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
 
-app.token.secret=5MSfAxtWOwgLPKcYr5Le6mKlv56cGK2gwZEAfWudaT5BAF_9cyS2lz3FxNjMpJ
-app.token.expiresIn=360
+application.token.secret=5MSfAxtWOwgLPKcYr5Le6mKlv56cGK2gwZEAfWudaT5BAF_9cyS2lz3FxNjMpJ
+application.token.expiresIn=360
